tests: Added unit test for card use-cases

+++ api/src/application/use-cases/card/DeleteCardUseCase.js
@@ -4,7 +4,17 @@ class DeleteCardUseCase {
   }
 
   async execute(cardId) {
-    return this.cardRepository.delete(cardId);
+    if (!cardId) throw new Error('cardId was not provided');
+
+    const result = await this.cardRepository.delete(cardId);
+
+    if (result === 0) {
+      throw new Error(
+        'Something went wrong deleting the card. Maybe the cardId does not exist',
+      );
+    }
+
+    return result;
   }
 }
